feat: Webhooks support (#16)
* feat: add message passing system for event publishing

- Add Event and MessagePublisher interface with MessagePublisherManager
- Add event type enum (internal/datatypes/event_type.go) for feedback_record
  and future webhook event types
- Add placeholder EmailDeliveryService implementing MessagePublisher
- Wire FeedbackRecordsService to publish events on create/update/delete
- Wire message manager in main and integration tests (no webhook provider yet)


* chore: removed excessive comments

* chore: remove PublishEvents

* chore: refactored service

* chore: further refactoring

* chore: webhooks implementation

* chore: implemented retryable http

* chore: concurrency limit; singleflight

* feat: goose based migrations

* Merge feat/goose-migrations into feat/webhooks

- Use goose for app schema (migrations/001, 002_webhooks)
- Single 002_webhooks.sql migration with full webhooks table and indexes
- Keep river-migrate target and add to CI after init-db
- AGENTS.md: document river-migrate for webhook delivery


* chore: changed context

* chore: updated openapi.yaml

* chore: fixed tenant_id

* chore: webhook id idempotency

* chore: remove file

* chore: changed webhooks redirect / timeouts to fall in line with standard webhooks

* chore: further alignment with standard webhooks

* chore: migration change / comments

* chore: slight event refactor

* chore: added publisher queue

* chore: message publisher worker shutdown

* chore: river based webhooks implementation

* chore: Updated linting rules, and fixed problems

* chore: additional performance and safety lint rules

* chore: additional performance linting rules

* chore: added riverui to docker compose

* Merge branch 'chore/linting-rules' into feat/webhooks


* chore: improve river performance by using batches ("InsertMany" instead of "Insert")

* chore: additional env vars in example

* chore: assorted fixes

* chore: messages on bulk deletes

* chore: vulnerability fix

* chore: updated deletes

* chore: fixed error checking

* chore: unnecessary error checking

* chore: url validation

* chore: add hard limit to webhooks

* chore: .env.example update

* chore: updated timestamp types in DB

* chore: fixed timestamps

* chore: early webhook key validation

* chore: remove limit in listing webhooks that listen for events

* chore: renamed for clarity

* chore: refactor changed fields

* chore: refactor main

* chore: merged migrations; updated indexes

* chore: refactored event types

* chore: fix integration tests

* chore: fix make

* chore: fix build

* chore: fix documentation

* chore: fix workflow

* chore: fix shutdown double call

---------

+++ internal/api/handlers/feedback_records_handler.go
@@ -23,7 +23,7 @@ type FeedbackRecordsService interface {
 	ListFeedbackRecords(ctx context.Context, filters *models.ListFeedbackRecordsFilters) (*models.ListFeedbackRecordsResponse, error)
 	UpdateFeedbackRecord(ctx context.Context, id uuid.UUID, req *models.UpdateFeedbackRecordRequest) (*models.FeedbackRecord, error)
 	DeleteFeedbackRecord(ctx context.Context, id uuid.UUID) error
-	BulkDeleteFeedbackRecords(ctx context.Context, userIdentifier string, tenantID *string) (int64, error)
+	BulkDeleteFeedbackRecords(ctx context.Context, userIdentifier string, tenantID *string) (int, error)
 }
 
 // FeedbackRecordsHandler handles HTTP requests for feedback records.
@@ -195,7 +195,7 @@ func (h *FeedbackRecordsHandler) BulkDelete(w http.ResponseWriter, r *http.Reque
 	}
 
 	resp := models.BulkDeleteResponse{
-		DeletedCount: deletedCount,
+		DeletedCount: int64(deletedCount),
 		Message:      fmt.Sprintf("Successfully deleted %d feedback records", deletedCount),
 	}
